[refactor] project 모듈을 filter로 이름 변경
- 기존 Project 모듈에서 모든 filter 데이터 관리를 위하여 filter로 이름 변경
